+++ danode/response.d
@@ -72,7 +72,7 @@ struct Response {
   @property final StatusCode statuscode() const { return payload.statuscode; }
   @property final bool keepalive() const { return( toLower(connection) == "keep-alive"); }
   @property final long length(){ if(payload.length >= 0){ return header.length + payload.length; }else{ return(long.max); } }
-  @property final const(char)[] bytes(in long maxsize = 1024){                              // Return the bytes from index to the end
+  @property final const(char)[] bytes(in ptrdiff_t maxsize = 1024){                              // Return the bytes from index to the end
     ptrdiff_t hsize = header.length;
     if(index <= hsize) return(header[index .. hsize] ~ payload.bytes(0, maxsize-hsize));    // We haven't completed the header yet
     return(payload.bytes(index-hsize));                                                     // Header completed, just stream bytes from the payload
